+++ src/app/(app)/lists/[id]/addButton/AddNamePicker.tsx
@@ -1,14 +1,11 @@
 import { Modal } from "app/(app)/shared/Modal";
-import { useAuth } from "app/login/authContext";
+import { useAuth } from "app/authContext";
+import { createListData, createSectionData } from "db/factory";
+import { lists, sectionsOfList } from "db/useDb";
 import { addDoc } from "firebase/firestore";
 import { useRouter } from "next/navigation";
 import { Dispatch, SetStateAction, useState } from "react";
 import { AdminList } from "types/types";
-import {
-  createListData,
-  createSectionData,
-} from "../../../../../firebase/factory";
-import { lists, sectionsOfList } from "../../../../../firebase/useDb";
 
 const AddNamePicker = ({
   activeList,
