+++ src/app/(app)/lists/[id]/items/item.module.css
@@ -51,6 +51,7 @@
   z-index: 0;
   border-radius: 3px;
   -webkit-appearance: none;
+  appearance: none;
   cursor: pointer;
 }
