+++ src/app/(app)/lists/[id]/items/items.module.css
@@ -1,8 +1,3 @@
-/* @media (max-width: 500px) {
-  .itemsArea {
-  }
-} */
-
 .itemsHeader {
   display: flex;
   gap: 15px;
@@ -26,7 +21,3 @@
 .itemList {
   margin-bottom: 10px;
 }
-
-/* .completedTodoHeader {
-  display: block;
-} */
