--- src/app/(app)/lists/[id]/lists/globals.css
@@ -1,130 +1,0 @@
-@import url("reset.css");
-
-html,
-body {
-  height: 100%;
-}
-
-body {
-  background: #202124;
-  color: #e8eaed;
-  font-family: sans-serif;
-  font-size: 16px;
-}
-
-#__next {
-  max-width: 960px;
-  margin: 0 auto;
-  height: 100%;
-}
-
-:global(#firebaseui-auth-container) {
-  grid-column-end: -1;
-  grid-column-start: 1;
-}
-
-.hiddenIcon {
-  display: none;
-}
-
-.hiddenIcon.active,
-.active .hiddenIcon {
-  display: initial;
-}
-
-[hidden] {
-  display: none !important;
-}
-
-h1 {
-  font-size: 24px;
-  font-weight: 700;
-  padding: 10px 0;
-  word-break: break-word;
-}
-h2 {
-  font-size: 20px;
-  font-weight: 500;
-  padding: 10px 0;
-  word-break: break-word;
-}
-
-button {
-  border: 1px solid rgb(245, 222, 178);
-  border-radius: 4px;
-  line-height: 25px;
-  margin: 10px;
-  padding: 0 5px;
-  cursor: pointer;
-}
-
-button:disabled {
-  border-color: rgba(255, 255, 255, 0.3);
-  cursor: default;
-}
-
-button:hover {
-  background: rgb(255 255 255 / 4%);
-}
-
-button.primary {
-  background: rgb(245, 222, 178);
-  color: black;
-}
-
-button.primary:hover {
-  background: rgb(251, 242, 223);
-}
-
-.md-inactive {
-  color: rgba(255, 255, 255, 0.3);
-}
-.icon {
-  cursor: pointer;
-  /* margin-left: 7px;
-  line-height: 100%; */
-}
-
-/* Rules for sizing the icon. */
-.material-icons.md-18 {
-  font-size: 18px;
-}
-.material-icons.md-24 {
-  font-size: 24px;
-}
-.material-icons.md-36 {
-  font-size: 36px;
-}
-.material-icons.md-48 {
-  font-size: 48px;
-}
-
-.yellow {
-  color: yellow;
-}
-
-.red {
-  color: red;
-}
-
-.teal {
-  color: lightskyblue;
-}
-
-.orange {
-  color: #fb8c00;
-}
-
-.green {
-  color: #b0fb00;
-}
-
-/* @media (prefers-color-scheme: dark) {
-  html {
-    color-scheme: dark;
-  }
-  body {
-    color: white;
-    background: black;
-  }
-} */
