+++ src/app/(app)/lists/page.tsx
@@ -1,6 +1,6 @@
 import { redirect } from "next/navigation";
 import { adminDb } from "../../../firebase/firebaseAdmin";
-import { verifyAuthToken } from "../../login/verifyAuthToken";
+import { verifyAuthToken } from "../../../auth/verifyAuthToken";
 
 const getFirstListId = async () => {
   const { user } = await verifyAuthToken();
